requirements.txt
angler==0.0.15
autograd==1.7.0
matplotlib==3.9.2
mlflow==2.17.0
mlflow_skinny==2.17.0
numpy==2.1.3
pyMKL==0.0.3
pyMKL==0.0.3
scipy==1.14.1
setuptools==75.1.0
skimage==0.0
torch==2.4.1
torchonn_pyutils==0.0.3.1
torchvision==0.19.1
